chain add-edit

+++ pfe-system-domain/src/main/java/dz/ade/pfe/domain/ouvrage/OuvrageChain.java
@@ -27,7 +27,7 @@ public class OuvrageChain extends Auditing{
     private Ouvrage ouvrage;
 
     @Column(name = "position")
-    private double position;
+    private Integer position;
 
 }
